# perf: speed up gossip node tests