## Motivation

Speed up node tests locally from 22s -> 10s by eliminating unnecessary waits and re-initializations. 

## Change Summary

- Two tests were combined to eliminate unnecessary double initialization
- Sleep times were reduced from 5s to2s since the network formed quickly
- `forEach` was replaced with `map` where iteration was not necessary

## Merge Checklist

- [x] The title of this PR adheres to the [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/) standard
- [x] The PR has been tagged with the appropriate change type label(s) (i.e. documentation, feature, bugfix, or chore)